backend/api_foodgram/requirements.txt
Django==3.2.16
django-filter==22.1
djangorestframework==3.14.0
reportlab==3.6.12
python-dotenv==0.21.0
django-cors-headers==3.13.0
djoser==2.1.0
django-import-export==3.0.1
django-colorfield==0.7.2
gunicorn==20.0.4
psycopg2-binary==2.8.6